added views

+++ src/Concerns/FileActions.php
@@ -2,13 +2,15 @@
 
 namespace Andresdevr\Administratix\Concerns;
 
+use Andresdevr\Administratix\Exceptions\FileDoesNotExists;
 use Andresdevr\Administratix\Facades\ConsoleOutput;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 
 trait FileActions
 {
     /**
-     * The clock
+     * The clocks array
      *
      * @var array
      */
@@ -21,11 +23,17 @@ trait FileActions
      * throw an error, but if has the option, create the file
      * 
      * @param  string $filename
+     * @param  string $reportIfNotExists
      * @return bool
      */
-    protected function fileExists($filename, $createIfNotExists = false)
+    protected function fileExists($filename, $reportIfNotExists = true)
     {
-
+        if(File::exists($filename)) 
+            return true;
+        if($reportIfNotExists)
+            throw new FileDoesNotExists($filename);
+        
+        return false;
     }
     
     /**
@@ -36,29 +44,55 @@ trait FileActions
      * @param  string $replacers
      * @return void
      */
-    public function replaceFile($from, $to = null, $replacers = [])
+    public function replaceFile($from, $to, $replacers = [], $reportIfNotExists = true)
     {
-        $this->startCommand("Replacing", $to);
+        $this->startClock("Replacing", $to);
 
-        $content = Str::of(file_get_contents($from));
+        if($this->fileExists($from, $reportIfNotExists))
+        {
+            $content = Str::of(file_get_contents($from));
+    
+            foreach($replacers as $search => $replace)
+                $content = $content->replace($search, $replace);
+    
+            file_put_contents($to, (string) $content);
+        }
 
-        foreach($replacers as $search => $replace)
-            $content = $content->replace($search, $replace);
-
-        file_put_contents($to, (string) $content);
-
-        $this->endCommand("Replaced", $to);
+        $this->endClock("Replaced", $to);
     }
 
     /**
      * Delete a file 
      * 
      * @param  string $filepathname
+     * @param  string $reportIfNotExists
      * @return bool
      */
-    public function deleteFile($filepathname)
+    public function deleteFile($filepathname, $reportIfNotExists = true)
     {
+        $this->startClock("Deleting", $filepathname);
 
+        if($this->fileExists($filepathname, $reportIfNotExists))
+            File::delete($filepathname);
+
+        $this->endClock("Deleted", $filepathname);
+    }
+
+    /**
+     * Copy a directory
+     * 
+     * @param  string $from
+     * @param  string $reportIfNotExists
+     * @param  string $to
+     */
+    public function copyDirectory($from, $to, $reportIfNotExists = true)
+    {
+        $this->startClock("Copying", $from);
+
+        if($this->fileExists($from, $reportIfNotExists))
+            File::copyDirectory($from, $to);
+
+        $this->endClock("Copied", $from);
     }
 
     /**
@@ -74,7 +108,7 @@ trait FileActions
             $this->writeln("<comment>{$action}:</comment>  {$file}") :
             ConsoleOutput::writeln("<comment>{$action}:</comment>  {$file}");
 
-        $this->clock[$action] = microtime(true);
+        $this->clock[$file] = microtime(true);
     }
 
     /**
@@ -84,7 +118,7 @@ trait FileActions
      */
     private function endClock($action, $file)
     {
-        $runTime = number_format((microtime(true) - $this->clock[$action]) * 1000, 2);
+        $runTime = number_format((microtime(true) - $this->clock[$file]) * 1000, 2);
 
         method_exists($this, 'getOutput') ?
             $this->writeln("<info>{$action}:</info>   {$file} ({$runTime}ms)") :
